Local benchmarks show good improvements for the crc32c intrinsification:


without intrinsic (master):

```
$JDK/java -DmsgSize=5120 -XX:TieredStopAtLevel=1 -Xcomp TestCRC32C 300000
 offset = 0
msgSize = 5120 bytes
  iters = 300000
-------------------------------------------------------
CRCs: crc = 0cbca9c8, crcReference = 0cbca9c8
CRC32C.update(byte[]) runtime = 1.186507782 seconds
CRC32C.update(byte[]) throughput = 1294.5553525244388 MB/s
CRCs: crc = 0cbca9c8, crcReference = 0cbca9c8
-------------------------------------------------------
CRCs: crc = 0cbca9c8, crcReference = 0cbca9c8
CRC32C.update(ByteBuffer) runtime = 1.355515648 seconds
CRC32C.update(ByteBuffer) throughput = 1133.1481139788364 MB/s
CRCs: crc = 0cbca9c8, crcReference = 0cbca9c8
-------------------------------------------------------
```

with intrinsic:

```
$JDK/java -DmsgSize=5120 -XX:TieredStopAtLevel=1 -Xcomp TestCRC32C 300000
 offset = 0
msgSize = 5120 bytes
  iters = 300000
-------------------------------------------------------
CRCs: crc = 0cbca9c8, crcReference = 0cbca9c8
CRC32C.update(byte[]) runtime = 0.065003188 seconds
CRC32C.update(byte[]) throughput = 23629.610289267657 MB/s
CRCs: crc = 0cbca9c8, crcReference = 0cbca9c8
-------------------------------------------------------
CRCs: crc = 0cbca9c8, crcReference = 0cbca9c8
CRC32C.update(ByteBuffer) runtime = 0.072310133 seconds
CRC32C.update(ByteBuffer) throughput = 21241.836189127185 MB/s
CRCs: crc = 0cbca9c8, crcReference = 0cbca9c8
-------------------------------------------------------
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8350866](https://bugs.openjdk.org/browse/JDK-8350866): [x86] Add C1 intrinsics for CRC32-C (**Enhancement** - P4)


### Reviewers
 * [Martin Doerr](https://openjdk.org/census#mdoerr) (@TheRealMDoerr - **Reviewer**)
 * [Vladimir Kozlov](https://openjdk.org/census#kvn) (@vnkozlov - **Reviewer**)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/23826/head:pull/23826` \
`$ git checkout pull/23826`

Update a local copy of the PR: \
`$ git checkout pull/23826` \
`$ git pull https://git.openjdk.org/jdk.git pull/23826/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 23826`

View PR using the GUI difftool: \
`$ git pr show -t 23826`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/23826.diff">https://git.openjdk.org/jdk/pull/23826.diff</a>

</details>
<details><summary>Using Webrev</summary>

[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/23826#issuecomment-2702162313)
</details>